Another attempt at #57 but not functional.
Attempted to replace the throwable requirement to be a 'perk_deck_unlock' instead of 'perkDeck', the intent being that a Copycat Mimicry can override perk_deck_unlock to look like a different perk, whilst other perk decks show the same perk_deck_unlock.

However, not sure why but the requirements unlocking thing doesn't seem to be letting me use 'perk_deck_unlock' as a requirement, which is a bit frustrating.